requirements.txt
requests~=2.32.3
python-dotenv~=0.21.0
pandas~=2.2.2
numpy~=1.26.4
loguru~=0.7.3
openai~=1.64.0
pyperclip~=1.9.0
selenium~=4.29.0
pyautogui~=0.9.54
web3~=7.9.0